refactor: move profiling to its own module
