+++ cyberdrop_dl/env.py
@@ -7,4 +7,5 @@ if KEY:
     KEY = sha256(KEY.encode("utf-8")).hexdigest()
 
 DEBUG_LOG_FILE_FOLDER = os.getenv("CDL_DEBUG_LOG_FILE_FOLDER")
-DEBUG_VAR = RUNNING_IN_IDE or DEBUG_LOG_FILE_FOLDER
+PROFILING = os.getenv("CDL_PROFILING")
+DEBUG_VAR = RUNNING_IN_IDE or DEBUG_LOG_FILE_FOLDER or PROFILING
